Adam branch

+++ ShoppingCartOOPSummativeAssessment/src/org/example/cart/Cart.java
@@ -2,9 +2,11 @@ package org.example.cart;
 
 import org.example.model.Item;
 
+import java.math.BigDecimal;
+
 public interface Cart {
     public void addItemToCart(Item item, int quantity);
     public void removeItemFromCart(Item item, int quantity);
-    public double getSubtotal();
+    public BigDecimal getSubtotal();
     public void checkout();
 }
